Add new lineshapes and automatic shape selection
This commit introduces new lineshapes: `no_apod` for no apodization, and `sp1` and `sp2` for shifted sine-bell apodization raised to the power of 1 and 2, respectively. The acquisition time and apodization parameters are used to accurately simulate the lineshape. By default, PeakFit will check the NMRPipe spectrum header to determine the appropriate model. If `SP` with powers of 1 or 2 is detected, `sp1` or `sp2` will be selected. If no apodization or `EM` is used, `no_apod` will be applied. For all other cases, `pvoigt` will be used.

Additionally, the segmentation algorithm for identifying clusters of peaks has been replaced by the SciPy algorithm.