feat: Add plotting module for peakfit (#5)
This commit adds a new module, `plotting`, to the `peakfit` package. It includes submodules for different types of plots, such as intensity and CEST plots. The `plotting` module provides functions for creating and saving plots based on the data generated by the `peakfit` package. This enhancement improves the visualization capabilities of the `peakfit` tool, allowing users to easily analyze and interpret the results of their spectral analysis.